+++ apx_core/src/crypto/mod.rs
@@ -1,0 +1,3 @@
+//! # Cryptographic primitives
+
+pub mod common;