+++ apx_core/src/lib.rs
@@ -5,6 +5,7 @@ pub mod base64;
 pub mod caip10;
 pub mod caip19;
 pub mod caip2;
+pub mod crypto;
 pub mod crypto_ecdsa;
 pub mod crypto_eddsa;
 pub mod crypto_rsa;